refactor: 상품서비스 커스텀 예외 처리 적용

+++ backend/src/main/java/com/team5/backend/domain/product/service/ProductService.java
@@ -14,6 +14,8 @@ import com.team5.backend.domain.product.dto.ProductResDto;
 import com.team5.backend.domain.product.dto.ProductUpdateReqDto;
 import com.team5.backend.domain.product.entity.Product;
 import com.team5.backend.domain.product.repository.ProductRepository;
+import com.team5.backend.global.exception.CustomException;
+import com.team5.backend.global.exception.code.ProductErrorCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -26,7 +28,7 @@ public class ProductService {
 
     public ProductResDto createProduct(ProductCreateReqDto request) {
         Category category = categoryRepository.findById(request.getCategoryId())
-                .orElseThrow(() -> new IllegalArgumentException("카테고리를 찾을 수 없습니다."));
+                .orElseThrow(() -> new CustomException(ProductErrorCode.CATEGORY_NOT_FOUND));
 
         Product product = Product.builder()
                 .category(category)
@@ -55,14 +57,13 @@ public class ProductService {
 
     public ProductResDto getProductById(Long productId) {
         Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다. ID: " + productId));
-
+                .orElseThrow(() -> new CustomException(ProductErrorCode.PRODUCT_NOT_FOUND));
         return ProductResDto.fromEntity(product);
     }
 
     public ProductResDto updateProduct(Long productId, ProductUpdateReqDto request) {
         Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new IllegalArgumentException("상품을 찾을 수 없습니다. ID: " + productId));
+                .orElseThrow(() -> new CustomException(ProductErrorCode.PRODUCT_NOT_FOUND));
 
         product.setTitle(request.getTitle());
         product.setDescription(request.getDescription());
@@ -74,6 +75,9 @@ public class ProductService {
     }
 
     public void deleteProduct(Long productId) {
+        if (!productRepository.existsById(productId)) {
+            throw new CustomException(ProductErrorCode.PRODUCT_NOT_FOUND);
+        }
         productRepository.deleteById(productId);
     }
 }
